% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{addDelta}
\alias{addDelta}
\title{Calculate the difference between expected and calculated log(m2cap)
for the last time period with floorspace data.}
\usage{
addDelta(data, lastYearWithData)
}
\arguments{
\item{data}{data frame containing historic and estimated floorspace data}

\item{lastYearWithData}{numeric, time period from which to compute the delta value}
}
\description{
Calculate the difference between expected and calculated log(m2cap)
for the last time period with floorspace data.
}
