% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildingsProjections.R
\name{addEURagg}
\alias{addEURagg}
\title{Aggregate various variables to european regions}
\usage{
addEURagg(df, extVars, intVars, floorVars, regionmap)
}
\arguments{
\item{df}{data to be aggregated}

\item{extVars}{variables of extensive items}

\item{intVars}{variables of intensive items}

\item{floorVars}{variables of floorspace items}

\item{regionmap}{data.frame that maps countries to regions}
}
\value{
data.frame with aggregated data on specific variables
}
\description{
Aggregate various variables to european regions
}
