% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildingsProjections.R
\name{buildingsProjections}
\alias{buildingsProjections}
\title{Make projections for useful and final energy levels at (enduse, carrier) level}
\usage{
buildingsProjections(
  config,
  floor,
  hddcdd,
  pop,
  gdppop,
  uvalue,
  pfu,
  feueEff,
  feSharesEC,
  regionmap,
  scenAssump,
  scenAssumpSpeed,
  scenAssumpCorrect,
  outputDir = "output",
  hpCorrection = TRUE,
  lifestyle = NULL,
  variableSpeed = NULL,
  ariadneFix = FALSE
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{floor}{historical and future residential/commercial floorspace (per capita)}

\item{hddcdd}{historical and future HDD's/CDD's}

\item{pop}{historical and future population data}

\item{gdppop}{historical and future gdp per capita}

\item{uvalue}{historical and future building insulation values}

\item{pfu}{historical final and useful energy data}

\item{feueEff}{historical and future FE->UE conversion efficiencies}

\item{feSharesEC}{historical and future final energy carrier shares}

\item{regionmap}{regional mapping}

\item{scenAssump}{carrier/enduse-specific scenario assumptions}

\item{scenAssumpSpeed}{long-term temporal scenario assumptions}

\item{scenAssumpCorrect}{specific corrections}

\item{outputDir}{output directory}

\item{hpCorrection}{heatpump correction for electric space heating}

\item{lifestyle}{lifestyle parameter}

\item{variableSpeed}{variable speed}

\item{ariadneFix}{ariadneFix flag}
}
\description{
This function covers the entire projection part of EDGE-B. It has been adapted
from the previous version of Antoine Levesque to allow scenario-wise processing
rather than running all scenarios at once.
}
\note{
The functionality of some features has not (yet) been integrated into this
new version. This includes the entire plot part, the correction for heat pumps,
\code{hpCorrection}, and the \code{lifestyle} and \code{variableSpeed} parameters.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
