% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariables.R
\name{calc_addRatio}
\alias{calc_addRatio}
\title{Calculate ratio}
\usage{
calc_addRatio(
  data,
  newNames,
  numerators,
  denominators,
  factor = NULL,
  units = NA,
  na.rm = TRUE,
  only.new = FALSE,
  variable = "variable",
  value = "value"
)
}
\arguments{
\item{data}{data.frame to work on}

\item{newNames}{variable name of the product}

\item{numerators}{variable to be the numerator}

\item{denominators}{variable of denominator}

\item{factor}{scaling factor}

\item{units}{character unit of calculated variable}

\item{na.rm}{logical, remove NAs}

\item{only.new}{logical, drop all variables but the newly calculated}

\item{variable}{character, name of variable column}

\item{value}{character, name of value column}
}
\value{
data.frame with ratio variable
}
\description{
This function calculates the ratio of two variables.
}
\author{
Antoine Levesque
}
