% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariables.R
\name{calc_addSubs}
\alias{calc_addSubs}
\title{Calculate Difference}
\usage{
calc_addSubs(
  data,
  newNames,
  vars2sumprod = NULL,
  factor = NULL,
  units = NA,
  na.rm = FALSE,
  completeMissing = TRUE,
  only.new = FALSE,
  variable = "variable",
  value = "value"
)
}
\arguments{
\item{data}{data.frame to work on}

\item{newNames}{variable name of the difference}

\item{vars2sumprod}{variables to calculate difference of}

\item{factor}{scaling factor}

\item{units}{character unit of calculated variable}

\item{na.rm}{logical, remove NAs}

\item{completeMissing}{logical, if \code{TRUE} fill missing values with zero}

\item{only.new}{logical, drop all variables but the newly calculated}

\item{variable}{character, name of variable column}

\item{value}{character, name of value column}
}
\value{
data.frame with difference variable
}
\description{
This function calculates the difference of variables.
}
\author{
Antoine Levesque
}
