% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariables.R
\name{calc_addSumProd}
\alias{calc_addSumProd}
\title{Calculate sum product}
\usage{
calc_addSumProd(
  data,
  newNames,
  vars2sumprod,
  collapse_sign,
  factor = NULL,
  units = NA,
  na.rm = FALSE,
  completeMissing = TRUE,
  only.new = FALSE,
  variable = "variable",
  value = "value"
)
}
\arguments{
\item{data}{data.frame to work on}

\item{newNames}{variable name of the sum product}

\item{vars2sumprod}{variables to calculate sum product of}

\item{collapse_sign}{operation between vars2sumprod, has to be "*" for sum
product}

\item{factor}{scaling factor}

\item{units}{character unit of calculated variable}

\item{na.rm}{logical, remove NAs}

\item{completeMissing}{logical, if \code{TRUE} fill missing values with zero}

\item{only.new}{logical, drop all variables but the newly calculated}

\item{variable}{character, name of variable column}

\item{value}{character, name of value column}
}
\value{
data.frame with product variable
}
\description{
This function calculates the sum product of two variables.
}
\author{
Antoine Levesque
}
