% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariables.R
\name{calcAddProd}
\alias{calcAddProd}
\title{Calculate product}
\usage{
calcAddProd(
  data,
  newNames,
  vars2sumprod = NULL,
  factor = NULL,
  units = NA,
  na.rm = TRUE,
  completeMissing = FALSE,
  only.new = FALSE,
  variable = "variable",
  value = "value"
)
}
\arguments{
\item{data}{data.frame to work on}

\item{newNames}{variable name of the product}

\item{vars2sumprod}{variables to calculate product of}

\item{factor}{scaling factor}

\item{units}{character unit of calculated variable}

\item{na.rm}{logical, remove NAs}

\item{completeMissing}{logical, if \code{TRUE} fill missing values with zero}

\item{only.new}{logical, drop all variables but the newly calculated}

\item{variable}{character, name of variable column}

\item{value}{character, name of value column}
}
\value{
data.frame with product variable
}
\description{
This function calculates the product of two variables.
}
\author{
Antoine Levesque
}
