% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{calcGlob}
\alias{calcGlob}
\title{Calculate global values of given variable(s)}
\usage{
calcGlob(df, vars2agg, nameGLO = "GLO", weights = NULL)
}
\arguments{
\item{df}{data containing values to be aggregated}

\item{vars2agg}{variables to globally aggregate}

\item{nameGLO}{name of global region}

\item{weights}{aggregation weights}
}
\value{
global values for given variables
}
\description{
Calculate global values of given variable(s)
}
\author{
Antoine Levesque
}
