% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLambda.R
\name{calcLambda}
\alias{calcLambda}
\title{Calculate time series of share parameter describing a linear, logistic, or
exponential transition}
\usage{
calcLambda(
  lastIdenticalYear,
  endYear,
  lambdaEnd,
  type = "logit",
  scaleLogit = 0.15,
  sharpnessExp = 1.8,
  startYearVector = 2005,
  step = 5
)
}
\arguments{
\item{lastIdenticalYear}{first year of transition}

\item{endYear}{last year of transition}

\item{lambdaEnd}{final value of share parameter (0 ... 1)}

\item{type}{interpolation function c('logit', 'linear', 'exp')}

\item{scaleLogit}{scale parameter of logistic function}

\item{sharpnessExp}{sharpness parameter for exponential function (higher = steeper initial rise)}

\item{startYearVector}{year from which to prepend yearly steps until
\code{lastIdenticalYear}}

\item{step}{numeric time step to determine for which years lambda is computed}
}
\value{
list with time series of share parameter
}
\description{
Calculate time series of share parameter describing a linear, logistic, or
exponential transition
}
\author{
Antoine Levesque, Hagen Tockhorn
}
