% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShares.R
\name{calcShares}
\alias{calcShares}
\title{Calculate share of values}
\usage{
calcShares(data, colShare, colVal = "value", ignoreColumns = NULL)
}
\arguments{
\item{data}{data.frame}

\item{colShare}{name of share column}

\item{colVal}{name of value column, default is "value"}

\item{ignoreColumns}{name of columns to ignore}
}
\description{
Calculates shares of elements in one column according to their value.
}
\author{
Antoine Levesque
}
