% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{capFloorProjections}
\alias{capFloorProjections}
\title{Correct floorspace projections with upper floorspace limit}
\usage{
capFloorProjections(df, capFloor, lambda, endOfHistory)
}
\arguments{
\item{df}{data.frame containing projected floorspace data}

\item{capFloor}{country-wise upper floorspace boundary}

\item{lambda}{temporal convergence shares}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\description{
Correct floorspace projections with upper floorspace limit
}
