% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkYaml.R
\name{checkYaml}
\alias{checkYaml}
\title{Check if .yaml file contains correct stores for targets pipeline}
\usage{
checkYaml()
}
\description{
Check if .yaml file contains correct stores for targets pipeline
}
\author{
Hagen Tockhorn
}
