% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLambda.R
\name{compLambdaScen}
\alias{compLambdaScen}
\title{Compute region-wise share of convergence from regionally differentiated
efficiencies towards long-term assumption for a specific scenario.}
\usage{
compLambdaScen(
  scenAssumpSpeed,
  varySpeed = 1,
  startYearVector = 2005,
  startPolicyYear = 2020
)
}
\arguments{
\item{scenAssumpSpeed}{years of full regional convergence for each scenario}

\item{varySpeed}{scale speed period of transition from 2015 onward
(e.g. full convergence in 2075 and \code{varySpeed = 2} leads to
full convergence in 2045)}

\item{startYearVector}{year from which to start time series (yearly steps)}

\item{startPolicyYear}{year from which to start convergence}
}
\value{
data.frame with region-wise convergence shares
}
\description{
Compute region-wise share of convergence from regionally differentiated
efficiencies towards long-term assumption for a specific scenario.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
