% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{completeHH}
\alias{completeHH}
\title{Complete historical data with countries not present in \code{dfPartial}}
\usage{
completeHH(dfPartial, dfComplete, endOfHistory)
}
\arguments{
\item{dfPartial}{historic data.frame on floorspace, population, floorspace per capita and population density}

\item{dfComplete}{data.frame with projected values}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\value{
complete data.frame
}
\description{
Fill data points and set future data/projection-deviation, \code{delta}, to
the last historical value.
}
