% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShareECprojections.R
\name{convergeFEShares}
\alias{convergeFEShares}
\title{Compute convergence towards target shares within a given time interval}
\usage{
convergeFEShares(
  feConverge,
  assumpFEShares,
  lambda,
  thist,
  tmax,
  fullConvergenceLevel,
  useCarrierInert,
  thresInert,
  useCarrierDh,
  thresDh
)
}
\arguments{
\item{feConverge}{data frame with historic shares, gdp and hdd}

\item{assumpFEShares}{data frame with FE target shares}

\item{lambda}{numeric, transition path for convergence}

\item{thist}{numeric, last historic year, i.e. the year from which convergence is computed}

\item{tmax}{numeric, last time period for which shares are to be projected}

\item{fullConvergenceLevel}{numeric, ratio of GDP to GDP in \code{thist}
at which full convergence should be reached}

\item{useCarrierInert}{character, end use-carrier combinations for which target shares
are halved if historic shares are low, i.e. below \code{thresInert}}

\item{thresInert}{numeric, threshold for historic shares below which target shares
of \code{useCarrierInert} are halved}

\item{useCarrierDh}{character, end use-carrier combinations for which the share remains constant
if heating degree days \code{hdd} are low, i.e. below \code{thresDh}}

\item{thresDh}{numeric, threshold for hdd below which target shares of \code{useCarrierDh}
are kept at historic levels}
}
\description{
Compute convergence towards target shares within a given time interval
}
