% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUvalues.R
\name{correctRegions}
\alias{correctRegions}
\title{Correct U-values for regional historical deviations}
\usage{
correctRegions(df, lambda, endOfHistory)
}
\arguments{
\item{df}{data.frame with historical and future U-values}

\item{lambda}{temporal convergence shares}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\value{
corrected U-values
}
\description{
Correct U-values for regional historical deviations
}
