% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.deriveElecHeatingAssump}
\alias{.deriveElecHeatingAssump}
\title{Calculate electric heating Asym from share and efficiency Asym}
\usage{
.deriveElecHeatingAssump(x)
}
\arguments{
\item{x}{data.frame with switches}
}
\value{
data.frame with additional electric heating Asym
}
\description{
Calculate electric heating Asym from share and efficiency Asym
}
