% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.expandToRegions}
\alias{.expandToRegions}
\title{Convert config to data frame with values for each region}
\usage{
.expandToRegions(config, regions, switches = NULL, prefix = NULL)
}
\arguments{
\item{config}{data.frame with one line per scenario and switches as columns}

\item{regions}{character vector of all regions}

\item{switches}{character vector of switch names (without prefix)}

\item{prefix}{common switch prefix that will be removed}
}
\value{
data.frame with switch values for each scenario and region
}
\description{
Convert config to data frame with values for each region
}
