% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProjections.R
\name{.fillHistoricalData}
\alias{.fillHistoricalData}
\title{Fill Historical Data Gaps Using Model Projections}
\usage{
.fillHistoricalData(
  data,
  estimate,
  var,
  periodBegin,
  endOfHistory,
  convReg = "absolute",
  convFactor = "lambda",
  transformVariable = NULL,
  transformVariableScen = NULL,
  acOwnershipRates = NULL
)
}
\arguments{
\item{data}{\code{data.frame} Input data frame containing historical data on}

\item{estimate}{\code{nls/lm} Fitted model object}

\item{var}{\code{character} Name of the target variable}

\item{periodBegin}{\code{numeric} first year of historical data}

\item{endOfHistory}{\code{numeric} Last year of historical data}

\item{convReg}{\code{character} Convergence type: "absolute" for value differences or "proportion" for ratios}

\item{convFactor}{\code{character} Convergence factor: "lambda" for slow linear convergence
or "fullconv" for fast logistic convergence}

\item{transformVariable}{\code{character} Variable transformation to apply (optional)}

\item{transformVariableScen}{\code{character} Scenario-specific variable transformation (optional)}

\item{acOwnershipRates}{\code{data.frame} AC penetration reference values}
}
\value{
\code{data.frame} Historical data with gaps filled using corrected projections
}
\description{
Fills gaps in the historical data of a single specified variable using model projections,
based on an estimate from a former regression. The data set must include the necessary variables
to ensure accurate estimation and correction.
}
\author{
Hagen Tockhorn
}
