% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.filterSwitches}
\alias{.filterSwitches}
\title{select switches and remove prefix}
\usage{
.filterSwitches(x, switches, prefix)
}
\arguments{
\item{x}{data.frame}

\item{switches}{character vector of switch names (without prefix)}

\item{prefix}{common switch prefix that will be removed}
}
\value{
data.frame with selected switches without prefix
}
\description{
select switches and remove prefix
}
