% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspace.R
\name{.prepScenInput}
\alias{.prepScenInput}
\title{Make a socio-economic driver compliant with the config file}
\usage{
.prepScenInput(df, cfgScen, scen)
}
\arguments{
\item{df}{data.frame with a scenario column}

\item{cfgScen}{scenario name to select from \code{df}}

\item{scen}{target scenario name to relabel to}
}
\description{
Make a socio-economic driver compliant with the config file
}
