% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.removePrefix}
\alias{.removePrefix}
\title{Remove prefix from character column}
\usage{
.removePrefix(x, prefix, col)
}
\arguments{
\item{x}{data.frame}

\item{prefix}{character, starting pattern to be removed}

\item{col}{character, column name}
}
\value{
data.frame where prefix has been removed from col
}
\description{
Remove prefix from character column
}
