% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.scenCol}
\alias{.scenCol}
\title{Turn row names into scenario column}
\usage{
.scenCol(x)
}
\arguments{
\item{x}{data.frame with row names}
}
\value{
data.frame with additional scenario column
}
\description{
Turn row names into scenario column
}
