% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProjections.R
\name{.transVar}
\alias{.transVar}
\title{Transform Variables in a Data Frame}
\usage{
.transVar(dfTrans, form, var)
}
\arguments{
\item{dfTrans}{\code{data.frame} A data frame containing the variable to be transformed.}

\item{form}{\code{character} A string representing the transformation formula.}

\item{var}{\code{character} The name of the variable to be transformed.}
}
\value{
\code{data.frame} A data frame with the transformed variable.
}
\description{
This function applies a transformation to a specified variable in a data frame.
}
