% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{.unpackDf}
\alias{.unpackDf}
\title{Unpack scenAssump entry to regional data.frame}
\usage{
.unpackDf(x, regions)
}
\arguments{
\item{x}{scenAssump value, either a data.frame or scalar number}

\item{regions}{character vector of all regions}
}
\value{
data.frame with region and value column
}
\description{
Unpack scenAssump entry to regional data.frame
}
