% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgebuildings-package.R
\docType{package}
\name{edgebuildings-package}
\alias{edgebuildings}
\alias{edgebuildings-package}
\title{edgebuildings: Model for the projection of global energy demand in the buildings sector}
\description{
The Energy Demand GEnerator projects energy demand for buildings both at the useful and final energy level. It covers the global demand and five energy services: space heating, space cooling, appliances and lighting (treated together) water heating and cooking. In a nutshell, it computes the useful energy intensity of each service in different regions, taking into account climate conditions as well as floor space demand. It then projects energy demand based on the assumption that, as developing countries grow economically, they will catch up with the energy service demand of developed countries. Further assumptions are made for the level beyond the current level of economic development (saturation or lower growth).
}
\author{
\strong{Maintainer}: Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})

Authors:
\itemize{
  \item Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})
  \item Antoine Levesque
  \item Hagen Tockhorn \email{hagento@pik-potsdam.de}
  \item Ricarda Rosemann \email{ricarda.rosemann@pik-potsdam.de} (\href{https://orcid.org/0009-0006-5939-3197}{ORCID})
  \item Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}
}

}
\keyword{internal}
