% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillHistoricalData.R
\name{fillHistoricalData}
\alias{fillHistoricalData}
\title{Fill Historical Data Gaps Using Model Predictions}
\usage{
fillHistoricalData(data, estimate, var, periodBegin, endOfHistory)
}
\arguments{
\item{data}{\code{data.frame} Input data frame containing historical data on}

\item{estimate}{\code{nls/lm} Fitted model object}

\item{var}{\code{character} Name of the target variable}

\item{periodBegin}{\code{numeric} first year of historical data}

\item{endOfHistory}{\code{numeric} Last year of historical data}
}
\value{
\code{data.frame} Historical data with gaps filled using corrected predictions
}
\description{
Fills gaps in the historical data of a single specified variable using model predictions,
based on an estimate from a former regression. The data set must include the necessary variables
to ensure accurate estimation and correction.
}
\author{
Hagen Tockhorn
}
