% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensity.R
\name{getDensity}
\alias{getDensity}
\title{Calculate population density}
\usage{
getDensity(fulltime = FALSE, pop, surface)
}
\arguments{
\item{fulltime}{boolean only return values until 2100 if TRUE}

\item{pop}{population data}

\item{surface}{surface area data}
}
\description{
Population density in countries in cap/km2.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
