% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEfficiencies.R
\name{getEfficiencies}
\alias{getEfficiencies}
\title{Projection of Final to Useful Energy Efficiencies}
\usage{
getEfficiencies(
  config,
  histEfficiencies,
  gdppop,
  scenAssump,
  scenAssumpSpeed,
  regPars = NULL,
  gasBioEquality = TRUE
)
}
\arguments{
\item{config}{\code{data.frame} Configuration dataframe specifying scenario settings}

\item{histEfficiencies}{\code{data.frame} Historical efficiency data}

\item{gdppop}{\code{data.frame} GDP per capita projections}

\item{scenAssump}{\code{data.frame} Scenario assumptions for projections}

\item{scenAssumpSpeed}{\code{data.frame} Convergence speed assumptions for scenarios}

\item{regPars}{\code{data.frame} Pre-calculated regression parameters (optional)}

\item{gasBioEquality}{\code{logical} Indicates whether gas and bioenergy should be treated equally}
}
\value{
\code{data.frame} Projected efficiencies for all carrier-enduse combinations
}
\description{
This function calculates the final (FE) to useful energy (UE) efficiencies for all
combinations of energy carriers and end-uses. The projections are based on a model
by De Stercke et al., primarily driven by GDP per capita, which follows an S-shaped
curve approaching assumed efficiency levels. The curve parameters are derived from a
regression using IEA data and are modified according to scenario assumptions to reflect
different narratives. The final projection is obtained via a linear interpolation
between the unmodified and modified projections.
}
\note{
Given that the regional resolution from the initial dataset (used to derive the
historical efficiencies) differs, we use the regression parameters obtained from
said regional resolution (\code{regPars}) in \code{makeProjections}.
}
\references{
De Stercke, S. (2014). Dynamics of Energy Systems: A Useful Perspective
(Interim Report, p. 68). IIASA. http://pure.iiasa.ac.at/id/eprint/11254/1/IR-14-013.pdf
}
\author{
Hagen Tockhorn
}
