% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFEUEefficiencies.R
\name{getFEUEefficiencies}
\alias{getFEUEefficiencies}
\title{Projection of final to useful energy efficiencies}
\usage{
getFEUEefficiencies(config, eff, gdppop, scenAssump, scenAssumpSpeed, regPars)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{eff}{data.frame with historic efficiencies}

\item{gdppop}{data.frame with historic and scenario gdp per capita}

\item{scenAssump}{region-wise scenario assumptions for regression parameters}

\item{scenAssumpSpeed}{scenario assumptions on temporal evolution}

\item{regPars}{regression parameters for enduse-carrier combinations}
}
\value{
data.frame with historic and scenario-projected fe->eu efficiencies
}
\description{
Calculate final (FE) to useful energy (UE) efficiencies for all combinations
of carrier and enduse. Projections are based on a model by De Stercke et al.
mainly driven by GDP per capita. It describes a S-shaped curve approaching
assumed efficiency levels. The parameters of that curve are derived by a
regression with observations of IEA data. The parameters are modified
according to scenario assumptions to reflect different narratives. The final
projection then follows from a linear interpolation between the results of
the unmodified and the modified projection.
}
\note{
Considering the rise in efficiency will reach a plateau, and the
heterogeneity of heat pumps, I take :
HP efficiency SSP1: 5, SSP2: 5, SSP3: 3.875, SSP4: a:3.875 b:5 c:5, SSP5: 6

Scenario assumptions for space_heating.elec
I take the same COPs as in space_cooling, and I had an assumption on the
penetration of heat pumps in space_heating.elec. I take an efficiency of 1
for electric heaters
SSP1 = 80\% -> 4.2 (1\emph{20\% + 80\%} 5)
SSP2 = 50\% -> 3
SSP3 = 25\% -> 1.71875
SSP4a = 25\% -> 1.71875
SSP4b = 40\% -> 2.6
SSP4c = 60\% -> 3.4
SSP5 = 75\% -> 4.75
for water heaters
SSP1 = 80\% -> 4.2 (1\emph{20\% + 80\%} 5)
SSP2 = 50\% -> 3
SSP3 = 25\% -> 1.71875
SSP4a = 25\% -> 1.71875
SSP4b = 40\% -> 2.6
SSP4c = 60\% -> 3.4
SSP5 = 75\% -> 4.75

For the initial share of heat pumps:
According to the IEA https://www.iea.org/tcep/buildings/heating/heatpumps/
Heat pumps meet app. 3\% of heating needs in buildings
From the ETP data at the global level, heating needs sum to 62 EJ (WH + SH)
Electriticity for WH and SH is amounts to 7 EJ
(0.03*62)/7 = app 25\%
The initial efficiency is therefore 0.75 + 0.25 * 3
I raise the scenario assumptions accordingly concerning the share of HP

the coefficients displayed on the plots correspond to the coefficients
of the function, as it is used by Simon De Stercke (that is why I take the
exponential of the negative of the coefficient of SSAsym). The maximum are
fairly close to the estimates of Simon, the minimum are not that far, and the
curvature parameter is very far. This is not very surprising since he uses
GDPppp $1990, and I $2005. Further, the gdppop used here differ from other
sources (like World Bank) which is annoying.
There are other reasons than gdppop for the differences observed and to
explain why the points here are not on a line. The regions are different than
in PFUDB. When we disaggregate to the ISO level, we make the assumption that
each country inside the group has the same efficiency, which means the same
income level, which is of course not the case. And when we reaggregate, we
this might make a difference. For appliances light, they mix refrigerators
(Low Heat), Light, stationary Power and others. So this is normal that they
do not correspond to any of these efficiencies.
It would be difficult to use their estimates directly because of the
different gdppop values, but this would be possible. I could justify redoing
the estimates to calibrate them to my gdppop data.
In their framework, there is actually no justification for a deviation from
the line. Here, they would only be justified by the composition of the
different efficiencies for appliances_light but not for the others. Still, I
would prefer not to rely on the estimates for appliances_light, and design on
own estimation here.

The notes above are outdated have not yet been updated nor checked
(HT, 03/2024)
}
\references{
De Stercke, S. (2014). Dynamics of Energy Systems: A Useful
Perspective (Interim Report, p. 68). IIASA.
http://pure.iiasa.ac.at/id/eprint/11254/1/IR-14-013.pdf

Scenario assumptions for space_cooling.elec: mostly based on
http://www.powerknot.com/how-efficient-is-your-air-conditioning-system.html
}
\author{
Antoine Levesque, Hagen Tockhorn
}
