% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceBuild.R
\name{getFloorspaceBuild}
\alias{getFloorspaceBuild}
\title{Project future floorspace demand}
\usage{
getFloorspaceBuild(
  config,
  resid,
  comShares,
  comSharesIEA,
  regionmappingIEA,
  regionmapping,
  scenAssumpSpeed
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{resid}{data on residential floorspace}

\item{comShares}{data on absolute share of commercial building floorspace demand}

\item{comSharesIEA}{data on absolute share of commercial building FE demand (IEA)}

\item{regionmappingIEA}{region mapping for IEA dataset}

\item{regionmapping}{region mapping}

\item{scenAssumpSpeed}{region-wise scenario assumptions on temporal evolution}
}
\value{
data.frame w/ absolute and type-wise floorspace in m2
}
\description{
Project floorspace demand of residential and commercial buildings from
historical data. This function aggregates the results from \code{getFloorspaceResidential}
and \code{getShareFloorCommercial}.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
