% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{getFloorspaceResidential}
\alias{getFloorspaceResidential}
\title{Residential floorspace demand projections}
\usage{
getFloorspaceResidential(
  config,
  floorspacePast,
  gdppop,
  density,
  pop,
  surface,
  regionmap,
  scenAssump,
  scenAssumpSpeed
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{floorspacePast}{data.frame historic floorspace demand}

\item{gdppop}{data.frame gdp per capita}

\item{density}{data.frame population density}

\item{pop}{data.frame population}

\item{surface}{data.frame with region surface area}

\item{regionmap}{regional mapping}

\item{scenAssump}{data.frame scenario-specific assumptions}

\item{scenAssumpSpeed}{data.frame scenario-specific temporal upper boundary}
}
\value{
data.frame historic and scenario-specific projected residential
floorspace in million m2 and floorspace per capita in m2/cap on regional level
}
\description{
Takes historic floorspace demand, income projections, population density
projections and population projections and projects residential floorspace
demand at country level.
}
\author{
Antoine Levesque, Hagen Tockhorn, Ricarda Rosemann
}
