% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGDP.R
\name{getGDP}
\alias{getGDP}
\title{Historic and projected GDP}
\usage{
getGDP(config, gdp, regionmap)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{gdp}{data.frame containing gdp data}

\item{regionmap}{regional mapping}
}
\description{
Adds GDP boost if specified in scenario config.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
