% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGDPpop.R
\name{getGDPpop}
\alias{getGDPpop}
\title{Historic and projected GDP}
\usage{
getGDPpop(pop, gdp, fulltime = FALSE)
}
\arguments{
\item{pop}{data.frame with population}

\item{gdp}{data.frame with GDP}

\item{fulltime}{data.frame logical, if \code{FALSE}, drop values after 2100}
}
\value{
data.frame with GDP per capita
}
\description{
Binds historic with scenario projections of population
}
\author{
Antoine Levesque, Hagen Tockhorn
}
