% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShareECprojections.R
\name{getShareECprojections}
\alias{getShareECprojections}
\title{Projected final energy share of carriers for each end use}
\usage{
getShareECprojections(
  config,
  pfu,
  hddcdd,
  gdp,
  gdppop,
  scenAssumpFEShares,
  regionalmap
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{pfu}{data.frame final and useful energy data}

\item{hddcdd}{historical and future HDD's/CDD's}

\item{gdp}{data.frame gdp data}

\item{gdppop}{data.frame gdp per capita}

\item{scenAssumpFEShares}{data.frame scenario-specific FE sare assumptions}

\item{regionalmap}{data.frame regional mapping}
}
\value{
data.frame
}
\description{
Project the final energy shares of carriers (EC) for each end use (EU) based
on assumed final shares.There are tuples of (EU, EC) that are phased out once
a particular threshold of GDP per capita is reached (energy ladder). The
remaining share is then split between the other carriers with their relative
shares approaching assumed shares. The speed of approaching final values is
determined by the maximum of a temporal and a GDP-driven convergence. The
temporal convergences is reached in 2150. The GDP-driven convergence is
reached, once GDP per capita is 40 times higher than it was at the end of
history (EOH).
}
\note{
for the variables without projections based on gdppop, use the
scenario assumptions from scenAssumpFEShares. Here we assume that the
share fully converge to the scenario assumption when the gdp reaches 10 times
the getOption("end_of_history") amount. This is an ad hoc assumption. The
shares depend on the development of the size of the economy. Further, a time
component has been introduced to allow rich countries to change their shares
as well. Time convergence by 2150.

Compute the projected rest: The gdppop based projections determined
what is not fueled with the projections (a - sum(projectShares)).
if biotrad cooking = 0.75, coal cooking = 0.02 and petrol cooking = 0.05,
the rest is 0.18
if the shares projected in projections_non.gdppop give natgas cooking = elec
cooking = 0.5, we will have natgas cooking = elec cooking = 0.5 * 0.18 = 0.09
}
\author{
Antoine Levesque, Hagen Tockhorn
}
