% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShareFloorCommercial.R
\name{getShareFloorCommercial}
\alias{getShareFloorCommercial}
\title{Compute share of commercial buildings}
\usage{
getShareFloorCommercial(
  config,
  subtype = "EDGE",
  gdppop,
  pop,
  floor0,
  regionalmap
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{subtype}{source c("EDGE", "IEA", "raw_pop")}

\item{gdppop}{data.frame of gdp per capita}

\item{pop}{data.frame of population}

\item{floor0}{data.frame with IEA floor space data}

\item{regionalmap}{data.frame with regional mapping}
}
\description{
Estimate share of commercial floor space for all countries by fitting a
SSgompertz model w.r.t. GDP per capita to IEA data.
}
\author{
Antoine Levesque, Hagen Tockhorn, Robin Hasse
}
