% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{history2SSPs}
\alias{history2SSPs}
\title{Split or merge scenario column w.r.t. historical periods}
\usage{
history2SSPs(df, reverse = FALSE, endOfHistory = NULL, scen = NULL)
}
\arguments{
\item{df}{data.frame to be split/merged}

\item{reverse}{boolean, if TRUE scenario is split into "history"/SSPs}

\item{endOfHistory}{upper temporal boundary of historical data}

\item{scen}{scenario}
}
\value{
merged or split-up data.frame
}
\description{
Split or merge scenario column w.r.t. historical periods
}
\author{
Antoine Levesque, Hagen Tockhorn
}
