% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUvalues.R
\name{implementScenAssump}
\alias{implementScenAssump}
\title{Implement scenario assumptions on U-value temporal evolution}
\usage{
implementScenAssump(tmp, scenAssumpFactor, lambda, minU, endOfHistory)
}
\arguments{
\item{tmp}{data.frame with historic and projected U-values}

\item{scenAssumpFactor}{region-wise scenario value assumptions}

\item{lambda}{temporal convergence shares}

\item{minU}{minimum U-value}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\value{
corrected future U-values
}
\description{
Implement scenario assumptions on U-value temporal evolution
}
