% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{joinReduceYearsRegion}
\alias{joinReduceYearsRegion}
\title{Bind dataframes with common regions and periods}
\usage{
joinReduceYearsRegion(df1, df2)
}
\arguments{
\item{df1}{data.frame to join}

\item{df2}{data.frame to join}
}
\description{
Bind those rows of two dataframes that correspond to common regions and
periods.
}
\author{
Antoine Levesque
}
