% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUvalues.R
\name{lineariseChange}
\alias{lineariseChange}
\title{Linearise U-value temporal evolution}
\usage{
lineariseChange(df, periodBegin, endOfHistory)
}
\arguments{
\item{df}{data.frame with U-value projections}

\item{periodBegin}{lower temporal boundary of historical data}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\value{
linearised U-values
}
\description{
Smoothing the temporal evolution of future U-values to neglect the impact
of variability of degree daysby doing a linear interpolation of uvalues in the
time period of \code{endOfHistory} and 2100.
}
