% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMadratData.R
\name{loadMadratData}
\alias{loadMadratData}
\title{Load input data from mredgebuildings}
\usage{
loadMadratData(inputdataRevision, forceDownload = FALSE)
}
\arguments{
\item{inputdataRevision}{character, revision of the EDGE-B input data}

\item{forceDownload}{boolean, TRUE will force downloading the input data}
}
\value{
path to madrat input data directory
}
\description{
Note: This function has been closely adapted from \code{loadMadratData} from BRICK
by Robin Hasse.
}
\author{
Hagen Tockhorn
}
