% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFEUEefficiencies.R
\name{makeEfficiencyProjections}
\alias{makeEfficiencyProjections}
\title{Project FE->EU efficiencies for different scenarios}
\usage{
makeEfficiencyProjections(df, var, regPars, scenAssump, lambda)
}
\arguments{
\item{df}{data.frame containing data on gdppop and historic efficiencies}

\item{var}{enduse-carrier combination as \code{var = <enduse>.<carrier>}}

\item{regPars}{regression parameters for enduse-carrier combination}

\item{scenAssump}{scenario assumptions for regression parameters}

\item{lambda}{temporal convergence shares}
}
\value{
data.frame containing historic and projected efficiencies
}
\description{
Efficiencies of a specific enduse-carrier combination are projected to an S-shaped
curve with gdppop as a free parameter.Pre-calculated regression parameters from
historic data in mredgebuildings are used for projecting future values.
Specific regression parameters can be adjusted ith the \code{scenAssump} parameters.
}
