% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHistory.R
\name{mergeHistory}
\alias{mergeHistory}
\title{Make the history scenario part of all other scenarios}
\usage{
mergeHistory(data)
}
\arguments{
\item{data}{either a list of data.frames, each with a history and one
projection scenario or one dataframe with a history and multiple projection
scenarios}
}
\description{
Make the history scenario part of all other scenarios
}
\author{
Robin Hasse
}
