% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{read scenario configuration}
\usage{
readConfig(
  config = getSystemFile("config", "configTest.csv", package = "edgebuildings"),
  default = getSystemFile("config", "default.csv", package = "edgebuildings"),
  regiongroups = getSystemFile("config", "region_groups.csv", package = "edgebuildings"),
  subtype = "all"
)
}
\arguments{
\item{config}{name of scenario config file}

\item{default}{name of default config file w/ historic and default scenario values}

\item{regiongroups}{region mapping for regional/economic groups}

\item{subtype}{specifies output}
}
\value{
data.frame with full scenario configuration
}
\description{
This function reads the scenario configuration file and creates a data frame
for targets to define dynamic branches.
}
\note{
TODO: allow child scenarios with this syntax: childScen(parentScen)
}
\author{
Robin Hasse, Hagen Tockhorn
}
