% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProjections.R
\name{readProjections}
\alias{readProjections}
\title{Read projections}
\usage{
readProjections(path, mergeHist = TRUE)
}
\arguments{
\item{path}{character, path to run folder}

\item{mergeHist}{logical, merge history with each projection scenario if TRUE}
}
\description{
Read all projection csv files in a given path and return them as one
data.frame
}
\author{
Robin Hasse
}
