% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariables.R
\name{spreadSomeVariables}
\alias{spreadSomeVariables}
\title{Spread some variables}
\usage{
spreadSomeVariables(df, vars2spread, colVar = "variable", valueCol = "value")
}
\arguments{
\item{df}{data.frame}

\item{vars2spread}{variables to be spread into}

\item{colVar}{column of variables}

\item{valueCol}{column of values}
}
\description{
Spread some variables
}
