% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScenInput.R
\name{buildScenInput}
\alias{buildScenInput}
\title{build EDGE-B function inputs from config file}
\usage{
buildScenInput(
  config,
  subtype = c("scen_assump", "scen_assump_speed", "fe_shares", "mapping"),
  regionmap,
  singleScen = TRUE,
  valueOnly = FALSE
)
}
\arguments{
\item{config}{data.frame name of scenario config df}

\item{subtype}{EDGE-B input type}

\item{regionmap}{regional mapping}

\item{singleScen}{single scenario only}

\item{valueOnly}{returns list of values with regions as row names}
}
\value{
data.frame with specified input data
}
\description{
This function reads the scenario configuration data frame from readConfig and
creates specified data frames as input for EDGE-B functions.
NOTE: This function is developed for a single scenario input.
}
\author{
Hagen Tockhorn, Robin Hasse
}
