% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLambda.R
\name{calcLambda}
\alias{calcLambda}
\title{Calculate time series of share parameter describing a linear or logistic
transition}
\usage{
calcLambda(
  startYear,
  endYear,
  lambdaEnd,
  type = "logit",
  scaleLogit = 0.15,
  startYearVector = 2005
)
}
\arguments{
\item{startYear}{first year of transition}

\item{endYear}{last year of transition}

\item{lambdaEnd}{final value of share parameter (0 ... 1)}

\item{type}{interpolation function c('logit', 'linear')}

\item{scaleLogit}{scale parameter of logistic function}

\item{startYearVector}{year from which to prepend yearly steps until
\code{startYear}}
}
\value{
list with time series of share parameter
}
\description{
Calculate time series of share parameter describing a linear or logistic
transition
}
\author{
Antoine Levesque
}
