% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{capFloorProjections}
\alias{capFloorProjections}
\title{Correct floorspace projections with upper floorspace limit}
\usage{
capFloorProjections(projectionData, capFloor, lambda, endOfHistory)
}
\arguments{
\item{projectionData}{data.frame containing projected floorspace data}

\item{capFloor}{country-wise upper floorspace boundary}

\item{lambda}{temporal convergence shares}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\description{
If already the last time point with historic data exceeds the cap, the floorspace
demand is gradually decreased towards the cap.
}
\details{
Otherwise the minimum of the projected floorspace demand and the cap is taken.
}
