% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUvalues.R
\name{getUvalues}
\alias{getUvalues}
\title{Historic and projected conductivity of building shell}
\usage{
getUvalues(
  config,
  uvaluesResCom,
  uvaluesETSAP,
  gdppop,
  hddcdd,
  pop,
  regionalmap,
  scenAssumpSpeed
)
}
\arguments{
\item{config}{scenario-wise parameter configuration}

\item{uvaluesResCom}{residential and commercial uvalue data}

\item{uvaluesETSAP}{uvalue data}

\item{gdppop}{gdp per capita}

\item{hddcdd}{hdd cdd data}

\item{pop}{population data}

\item{regionalmap}{regional mapping}

\item{scenAssumpSpeed}{scenario-specific upper temporal evolution boundary}
}
\value{
data.frame
}
\description{
Historic data for conductivity of building shell by the EU Buildings Observatory
and IEA ETSAP along with historic and projected data on GDP per capita for
various scenarios is used to calibrate a regression model to project the
scenario-wise evolution of conductivity values.
}
\details{
Average U-values are estimated as the arithmetic mean of floor, walls, windows
and roofs U-values, implying an assumption of equal surface of each of these
buildings' compomnents.

As an estimation model we assume an exponential relation between the sum of
annual degree days (HDD/CDD) and the U-value. As the wished level of
insulation might be unaffordable for low-income households/businesses, we multiply
the U-value computed by the previous equation by a coefficient decreasing from
two to one when the per capita income of the region considered moves from US$0 to US$15000.
}
\author{
Antoine Levesque, Hagen Tockhorn
}
