% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateICT.R
\name{integrateICT}
\alias{integrateICT}
\title{Integrate Exogenous ICT Demand Projections}
\usage{
integrateICT(ict, fe, config)
}
\arguments{
\item{ict}{\code{data.frame} Exogenous ICT electricity demand data containing
historical and scenario projections.}

\item{fe}{\code{data.frame} Historical final energy data}

\item{config}{\code{data.frame} Configuration dataframe specifying scenario settings}
}
\value{
\code{list} A list with two elements:
\item{fe}{Updated final energy data}
\item{feICT}{Complete ICT electricity demand data}
}
\description{
This function integrates exogenous ICT (Information and Communication Technology)
demand projections by subtracting historical ICT demand from appliances & lighting (A&L)
electricity demand where it was previously included. The function first calculates
historical ICT shares relative to total A&L electricity demand, then continues these
shares with constant values where data is missing. The resulting ICT demand is then
subtracted from A&L electricity to create a separate ICT demand category, while future
projections are taken directly from the exogenous ICT data according to the specified
scenario.
}
\author{
Hagen Tockhorn
}
