% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFloorspaceResidential.R
\name{predictFullDelta}
\alias{predictFullDelta}
\title{Calculate full residential floorspace prediction}
\usage{
predictFullDelta(
  fullData,
  coefLogGdpEstimate,
  coefLogDensityEstimate,
  interceptEstimate,
  scenAssump,
  lambda,
  endOfHistory
)
}
\arguments{
\item{fullData}{full dataset containing predicted floorspace values}

\item{coefLogGdpEstimate}{logarithmic income elasticity parameter}

\item{coefLogDensityEstimate}{logarithmic pop density elasticity parameter}

\item{interceptEstimate}{logarithmic intercept parameter}

\item{scenAssump}{parameter scenario assumptions}

\item{lambda}{temporal convergence factors}

\item{endOfHistory}{upper temporal boundary of historical data}
}
\description{
As the income elasticity will evolve over time according to scenario assumptions,
the elasticity at a certain point in time will only influence the incremental
floor space demand. This implies a stepwise calculation of future floor space demand
based on the value in the previous time step. The resulting iterative formula can be
rearranged to an equivalent explicit formulation of the following form:
Regression Model + Delta + growth correction term,
where the latter corrects the regression model for the fact that the changed income
elasticity only acts on income growth, i.e. this accounts for the original iterative approach.
}
