% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepHDDCDD.R
\name{prepHDDCDD}
\alias{prepHDDCDD}
\title{Pre-process degree day data for EDGE-B}
\usage{
prepHDDCDD(hddcdd, config, regionmap)
}
\arguments{
\item{hddcdd}{data.frame with degree days per year}

\item{config}{data.frame with configuration parameters for specific scenario}

\item{regionmap}{region mapping}
}
\value{
converged degree day data
}
\description{
Degree day data is filtered and processed according to specifications made
in the \code{config} file. Scenarios assume a gradual shift in limit temperatures
for both heating and cooling degree days relative to historic thresholds, i.e.
heating and cooling behavior changes w.r.t to time and assumed scenario.
}
\details{
Hence, degree day data converges from historic thresholds \code{tlimHist} to
assumed target threshold \code{tlimTarget} in a time period from \code{endOfHistory}
to \code{yearTarget}. The convergence is done region- and variable-wise.
}
\author{
Hagen Tockhorn
}
