% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResults.R
\name{reportResults}
\alias{reportResults}
\title{Create reporting file summarising EDGE-B runs}
\usage{
reportResults(path, reporting, cfgText)
}
\arguments{
\item{path}{character, path to run folder}

\item{reporting}{reporting configuration, either a path to a file or the name
of an internal reporting configuration}

\item{cfgText}{character, configuration description to be added to the output file}
}
\description{
Create reporting file summarising EDGE-B runs
}
\author{
Robin Hasse
}
