% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeBuildings.R
\name{runEdgeBuildings}
\alias{runEdgeBuildings}
\title{Run EDGE-Buildings}
\usage{
runEdgeBuildings(
  config = "config_remind.csv",
  outputDir = "./output",
  reporting = NULL,
  madratDir = NULL,
  inputdataRevision = "0.5.12",
  forceDownload = FALSE,
  scenario = NULL
)
}
\arguments{
\item{config}{scenario configuration. Can be a path to a config file or the
name of an internal config file.}

\item{outputDir}{path to a directory to write files to}

\item{reporting}{reporting to apply}

\item{madratDir}{path to madrat root folder}

\item{inputdataRevision}{character, revision of the EDGE-B input data}

\item{forceDownload}{boolean, TRUE will force downloading the input data}

\item{scenario}{character vector, subset of scenarios to run. If NULL (default),
all scenarios in the config file will be run. Scenario names correspond to
column names in the config file (e.g., c("SSP2", "SSP5")). Note: SSP2 is
always required and will be automatically included if not specified, as it
provides reference values and is used for carrier share extension between
historic data and scenario start.}
}
\description{
Run EDGE-Buildings
}
\author{
Falk Benke, Robin Hasse
}
