% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeBuildings.R
\name{runEdgeBuildings}
\alias{runEdgeBuildings}
\title{Run EDGE-Buildings}
\usage{
runEdgeBuildings(
  config = "configEDGEscens.csv",
  outputDir = "./output",
  reporting = NULL,
  madratDir = NULL,
  inputdataRevision = "0.5",
  forceDownload = FALSE
)
}
\arguments{
\item{config}{scenario configuration. Can be a path to a config file or the
name of an internal config file.}

\item{outputDir}{path to a directory to write files to}

\item{reporting}{reporting to apply}

\item{madratDir}{path to madrat root folder}

\item{inputdataRevision}{character, revision of the EDGE-B input data}

\item{forceDownload}{boolean, TRUE will force downloading the input data}
}
\description{
Run EDGE-Buildings
}
\author{
Falk Benke, Robin Hasse
}
