% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildingsProjections.R
\name{splitElec}
\alias{splitElec}
\title{Split heat pump and resistive electric fe and ue}
\usage{
splitElec(df, feueEff, enduseChar, scenAssump, endOfHistory)
}
\arguments{
\item{df}{data frame containing share and efficiency data}

\item{feueEff}{historical and future FE->UE conversion efficiencies}

\item{enduseChar}{character, enduse for which electric FE demand should be split}

\item{scenAssump}{carrier/enduse-specific scenario assumptions}

\item{endOfHistory}{Last historic time period}
}
\value{
data frame containing split fe and ue
}
\description{
Split heat pump and resistive electric fe and ue
}
