% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseScenarios.R
\name{visualiseScenarios}
\alias{visualiseScenarios}
\title{Visualise scenarios}
\usage{
visualiseScenarios(path, outputFile = NULL, recoverRegions = "DEU")
}
\arguments{
\item{path}{character, path to run folder or named vector of paths with
version tags as names}

\item{outputFile}{path character, path to pdf that is created. If NULL, the
file projection.pdf is created in the first path provided in \code{path}.}

\item{recoverRegions}{character, edgebuildings regions to be recovered from
aggregated regions in order to plot them separately.}
}
\description{
This function should be replaced by a mif reporting and mip plotting at some
point.
}
\author{
Robin Hasse
}
